taggerBot/requirements.txt
pyTelegramBotAPI==4.2.2
paho-mqtt==1.6.1
tornado==6.1
numpy==1.22.0
pandas==1.3.5
scikit-learn==1.0.2